# docs/swagger.yaml
basePath: /shyft
definitions:
  handlers.RespondJson:
    properties:
      intent:
        type: string
      message: {}
      status:
        type: boolean
    type: object
...
info:
  contact:
    email: YunusAlpu@icloud.com
    name: Yunus Emre Alpu
    url: https://yunusemrealpu.netlify.app
  title: Shift Scheduler Service API
  version: 1.0.0
paths:
...
  /shift-schedules/{id}:
    delete:
      consumes:
      - application/json
      description: delete a shift schedule
      parameters:
      - description: Shift Schedule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: successfully deleted shift schedule
          schema:
            $ref: '#/definitions/handlers.RespondJson'
        "400":
          description: cannot delete shift schedule due to invalid request body
          schema:
            $ref: '#/definitions/handlers.RespondJson'
        "422":
          description: cannot delete shift schedule due to invalid request body
          schema:
            $ref: '#/definitions/handlers.RespondJson'
        "500":
          description: cannot delete shift schedule due to internal server error
          schema:
            $ref: '#/definitions/handlers.RespondJson'
      security:
      - BearerAuth: []
      summary: delete a shift schedule
      tags:
      - Shift
  /shift-schedules/{id}/restore:
    patch:
      consumes:
      - application/json
      description: restored a shift schedule
      parameters:
      - description: Shift Schedule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: successfully restored shift schedule
          schema:
            $ref: '#/definitions/handlers.RespondJson'
        "400":
          description: cannot restore shift schedule due to invalid request body
          schema:
            $ref: '#/definitions/handlers.RespondJson'
        "422":
          description: cannot restore shift schedule due to invalid request body
          schema:
            $ref: '#/definitions/handlers.RespondJson'
        "500":
          description: cannot restore shift schedule due to internal server error
          schema:
            $ref: '#/definitions/handlers.RespondJson'
      security:
      - BearerAuth: []
      summary: restore a shift schedule
      tags:
      - Shift
  /shift-schedules/{year}:
    get:
      consumes:
      - application/json
      description: get a shift schedules by year
      parameters:
      - description: Shift Schedule Year
        in: path
        name: year
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: get shifts by year successfully
          schema:
            $ref: '#/definitions/handlers.RespondJson'
        "400":
          description: cannot get shifts schedule by year due to invalid request body
          schema:
            $ref: '#/definitions/handlers.RespondJson'
        "422":
          description: cannot get shifts schedule by year due to invalid request body
          schema:
            $ref: '#/definitions/handlers.RespondJson'
        "500":
          description: cannot get shifts schedule by year due to internal server error
          schema:
            $ref: '#/definitions/handlers.RespondJson'
      security:
      - BearerAuth: []
      summary: get a shift schedules by year
      tags:
      - Shift
...
  /shift-schedules/week:
    get:
      consumes:
      - application/json
      description: get shift schedules by current week
      produces:
      - application/json
      responses:
        "200":
          description: get shifts by current week successfully
          schema:
            $ref: '#/definitions/handlers.RespondJson'
        "400":
          description: cannot get shifts schedule by current week due to invalid request
            body
          schema:
            $ref: '#/definitions/handlers.RespondJson'
        "422":
          description: cannot get shifts schedule by current week due to invalid request
            body
          schema:
            $ref: '#/definitions/handlers.RespondJson'
        "500":
          description: cannot get shifts schedule by current week due to internal
            server error
          schema:
            $ref: '#/definitions/handlers.RespondJson'
      security:
      - BearerAuth: []
      summary: get shift schedules by current week
      tags:
      - Shift
  /shift-schedules/week/paginated:
    get:
      consumes:
      - application/json
      description: get shift schedules by week with pagination
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: size
        in: query
        name: size
        type: string
      - description: orderBy
        enum:
        - asc
        - desc
        in: query
        name: orderBy
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: get shifts by current week successfully
          schema:
            $ref: '#/definitions/handlers.RespondJson'
        "400":
          description: cannot get shifts schedule by current week due to invalid request
            body
          schema:
            $ref: '#/definitions/handlers.RespondJson'
        "422":
          description: cannot get shifts schedule by current week due to invalid request
            body
          schema:
            $ref: '#/definitions/handlers.RespondJson'
        "500":
          description: cannot get shifts schedule by current week due to internal
            server error
          schema:
            $ref: '#/definitions/handlers.RespondJson'
      security:
      - BearerAuth: []
      summary: get shift schedules by current week with pagination
      tags:
      - Shift
schemes:
- http
- https
swagger: "2.0"